refactor: Moving updatePlan to a server util
