+++ app/routes/api.admin.update-plan/route.tsx
@@ -1,11 +1,6 @@
-import { ActionFunction, json } from "@remix-run/node"
-import invariant from "tiny-invariant"
-import { initPortalClient } from "~/models/portal/portal.server"
-import { AdminUpdateAccountMutationVariables, PayPlanType } from "~/models/portal/sdk"
-import { initAdminPortal } from "~/utils/adminPortal"
-import { getErrorMessage } from "~/utils/catchError"
-import { triggerSubscriptionActionNotification } from "~/utils/notifications.server"
-import { getPlanName } from "~/utils/planUtils"
+import { ActionFunction } from "@remix-run/node"
+import { PayPlanType } from "~/models/portal/sdk"
+import { updatePlan } from "~/utils/updatePlan.server"
 
 export type UpdatePlanActionData = {
   error: boolean
@@ -32,51 +27,4 @@ export const action: ActionFunction = async ({ request }) => {
     limit,
     subscription,
   })
-}
-
-export const updatePlan = async ({
-  id,
-  type: planType,
-  limit,
-  subscription,
-}: UpdatePlanArgs) => {
-  const portal = initPortalClient()
-
-  try {
-    invariant(id, "account id not found")
-    invariant(planType, "plan type not found")
-
-    const portalAdmin = await initAdminPortal(portal)
-
-    const options: AdminUpdateAccountMutationVariables = {
-      input: { accountID: id, payPlanType: planType },
-    }
-
-    if (limit) {
-      options.input.enterpriseLimit = limit
-    }
-    if (subscription) {
-      options.input.stripeSubscriptionID = subscription
-    }
-
-    await portalAdmin.adminUpdateAccount(options)
-
-    if (planType !== PayPlanType.FreetierV0) {
-      await triggerSubscriptionActionNotification({
-        planType,
-        accountId: id,
-        type: "upgrade",
-      })
-    }
-
-    return json<UpdatePlanActionData>({
-      error: false,
-      message: `Account ${id} has successfully been updated to ${getPlanName(planType)}`,
-    })
-  } catch (error) {
-    return json<UpdatePlanActionData>({
-      error: true,
-      message: getErrorMessage(error),
-    })
-  }
 }
