+++ app/routes/api.stripe.subscription-delete/route.tsx
@@ -2,9 +2,9 @@ import { ActionFunction, json } from "@remix-run/node"
 import invariant from "tiny-invariant"
 import { PayPlanType } from "~/models/portal/sdk"
 import { getSubscription, stripe } from "~/models/stripe/stripe.server"
-import { updatePlan } from "~/routes/api.admin.update-plan/route"
 import { getErrorMessage } from "~/utils/catchError"
 import { triggerSubscriptionActionNotification } from "~/utils/notifications.server"
+import { updatePlan } from "~/utils/updatePlan.server"
 import { getPoktId, requireUser } from "~/utils/user.server"
 
 export const action: ActionFunction = async ({ request }) => {
